update README.md
Add more subtasks to accomplish to learn how to use Flask for
this project, as well as add more resources around evaluating
discards.

I used primarily Cribbage Classic when doing spot checks
with my EV calculation. It would be useful to use the `requests'
library to write a fully automated testing system for this
function (and potentially others in the future).