add end of game popup, fix board colors
Creates a popup once a player wins the game. Adds a text message
to send to others that I have not tested yet.

The colors per card were incorrect: it is supposed to be (for both
cards) red for 2 wrong, yellow for 1 wrong, green for 0 wrong. Instead,
was giving massive hints by labeling individual cards, making the
game trivial. Fix forces the player to know both cards, not just one

... it is possible to work this out via deduction, but the point is
the increased effort, not the exact result.